Finished writing and manual testing the PUT request for editing an activity #story[326]
Finished writing and manual testing the PUT request for editing an activity, may need to revisit the way it pulls the set of activity types from the jsonnode object. #story[326]

+++ server/src/main/java/com/springvuegradle/seng302team600/controller/ActivityController.java
@@ -1,9 +1,11 @@
 package com.springvuegradle.seng302team600.controller;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.springvuegradle.seng302team600.Utilities.ActivityValidator;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.springvuegradle.seng302team600.model.Activity;
+import com.springvuegradle.seng302team600.model.User;
 import com.springvuegradle.seng302team600.repository.ActivityRepository;
 import com.springvuegradle.seng302team600.service.ActivityTypeService;
 import com.springvuegradle.seng302team600.service.UserValidationService;
@@ -15,7 +17,10 @@ import org.springframework.web.server.ResponseStatusException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.time.LocalDateTime;
+import java.sql.Date;
 import java.util.List;
+import java.util.Set;
 
 /**
  * Controller to manage activities and activity type
@@ -67,58 +72,69 @@ public class ActivityController {
         return allActivities;
     }
 
+
+    /** Put Request for editing/updating an activity created by a user
+     *  Checks all possible inputs to see if that input is there, and to be updated
+     *  takes from the client only the json object of the to-be-updated inputs
+     *  and the activity id through put the url mapping.
+     */
     @PutMapping("/activities/{activityId}")
-    public void editActivity(@PathVariable Long activityId, HttpServletRequest request,
-                             HttpServletResponse response, @RequestBody String jsonActivityEditString) throws IOException {
+    public void editActivity(@PathVariable Long activityId, HttpServletRequest request, HttpServletResponse response,
+                             @RequestBody String jsonActivityEditString) throws IOException {
         String token = request.getHeader("Token"); //this is the users token
         Activity activity = activityRepository.findByActivityId(activityId);
-
         //ResponseStatusException thrown if user unauthorized or forbidden from accessing requested user
         ObjectMapper nodeMapper = new ObjectMapper();
         ObjectNode editedData = nodeMapper.readValue(jsonActivityEditString, ObjectNode.class);
-
         //Get the fields to be updated from editedData object
         String newDescription = editedData.get("description").asText();
         String newLocation = editedData.get("location").asText();
         String newName = editedData.get("activity_name").asText();
         String checkContinuous = editedData.get("continuous").asText();
-        Boolean newContinuous = editedData.get("continuous").asBoolean();
-        String newStartDate = editedData.get("start_date").asText();
-        String newEndDate = editedData.get("end_date").asText();
-        String newActivityTypes = editedData.get("activity_type").asText();
+        JsonNode nodeActivityTypes = editedData.get("activity_type");
+        //may need to re-write the activity types conversion if breaks
+        Set newActivityTypes = nodeMapper.convertValue(nodeActivityTypes, Set.class);
         //Check if any have changed or are null
         if (newDescription != null) { activity.setDescription(newDescription); }
         if (newLocation != null) {activity.setLocation(newLocation);}
         if (newName != null) { activity.setName(newName); }
-        if (checkContinuous != null) { activity.setContinuous(newContinuous);}
-//        if (newStartDate != null) { Date date = new Date(); date.getDate()newStartDate; activity.setStartTime(date);}
-//        if (newEndDate != null) {}
-//        if (newActivityTypes != null) {
-//            activity.setActivityTypes(
-//                activityTypeService.getMatchingEntitiesFromRepository(newActivityTypes)
-//            );
-//
-//        }
+        if (newActivityTypes != null) {
+            activity.setActivityTypes(
+                activityTypeService.getMatchingEntitiesFromRepository(newActivityTypes)
+            );
+        }
+        if (checkContinuous != null) {
+            Boolean newContinuous = editedData.get("continuous").asBoolean();
+            if (checkContinuous != null) { activity.setContinuous(newContinuous);}
+            if (newContinuous == false) {
+                String newStartTime = editedData.get("start_time").asText();
+                String newEndTime = editedData.get("end_time").asText();
+                activity.setStartTime(Date.valueOf(newStartTime));
+                activity.setEndTime(Date.valueOf(newEndTime)); //This date manipulation relies on import java.sql.Date
+            }
+        }
+        activityRepository.save(activity);
+        response.setStatus(HttpServletResponse.SC_OK); //200
     }
 
     @DeleteMapping("/activities/{activityId}")
     public void deleteActivity(@PathVariable Long activityId, HttpServletRequest request) {
         //Check the activity is this specific users activity
         Activity activity = activityRepository.findByActivityId(activityId);
-//        Long authorId = activity.getCreatorUserId();
-//        String token = request.getHeader("Token");
-//        User user = userValidationService.findByToken(token); //finds user and validates they exist
-//        if (!userValidationService.hasAdminPrivileges(user)) {
-//            /* Only run this check if the user is NOT an Admin,
-//               otherwise admins can delete/edit others activities.
-//             */
-//            if (!authorId.equals(user.getUserId())) {
-//                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User is not activity creator");
-//            }
-//        }
-//        if (activity == null) {
-//            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Invalid activity id");
-//        }
+        Long authorId = activity.getCreatorUserId();
+        String token = request.getHeader("Token");
+        User user = userValidationService.findByToken(token); //finds user and validates they exist
+        if (!userValidationService.hasAdminPrivileges(user)) {
+            /* Only run this check if the user is NOT an Admin,
+               otherwise admins can delete/edit others activities.
+             */
+            if (!authorId.equals(user.getUserId())) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User is not activity creator");
+            }
+        }
+        if (activity == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Invalid activity id");
+        }
         //Delete the activity
         activityRepository.delete(activity);
